requirements.txt
asgiref==3.4.1
Django==4.0.1
django-adminlte-3==0.1.6
django-ckeditor==6.2.0
django-js-asset==1.2.2
filemagic==1.6
Pillow==9.0.0
sqlparse==0.4.2
tzdata==2021.5